Remove some duplications in the gui code.
