+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/screen/ControllerScreen.java
@@ -9,9 +9,7 @@ import com.refinedmods.refinedstorage2.platform.common.screen.widget.RedstoneMod
 import java.util.Collections;
 import java.util.List;
 
-import com.mojang.blaze3d.systems.RenderSystem;
 import com.mojang.blaze3d.vertex.PoseStack;
-import net.minecraft.client.renderer.GameRenderer;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.player.Inventory;
@@ -27,9 +25,6 @@ public class ControllerScreen extends AbstractBaseScreen<ControllerContainerMenu
     public ControllerScreen(final ControllerContainerMenu menu, final Inventory playerInventory, final Component text) {
         super(menu, playerInventory, text);
 
-        this.titleLabelX = 7;
-        this.titleLabelY = 7;
-        this.inventoryLabelX = 7;
         this.inventoryLabelY = 94;
         this.imageWidth = 176;
         this.imageHeight = 189;
@@ -55,6 +50,11 @@ public class ControllerScreen extends AbstractBaseScreen<ControllerContainerMenu
         ));
     }
 
+    @Override
+    protected ResourceLocation getTexture() {
+        return TEXTURE;
+    }
+
     private double getPercentageFull() {
         return (double) getMenu().getStored() / (double) getMenu().getCapacity();
     }
@@ -69,27 +69,8 @@ public class ControllerScreen extends AbstractBaseScreen<ControllerContainerMenu
     }
 
     @Override
-    protected void renderBg(final PoseStack poseStack, final float delta, final int mouseX, final int mouseY) {
-        RenderSystem.setShader(GameRenderer::getPositionTexShader);
-        RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, 1.0F);
-        RenderSystem.setShaderTexture(0, TEXTURE);
-
-        final int x = (width - imageWidth) / 2;
-        final int y = (height - imageHeight) / 2;
-
-        blit(poseStack, x, y, 0, 0, imageWidth, imageHeight);
-    }
-
-    @Override
     protected void renderLabels(final PoseStack poseStack, final int mouseX, final int mouseY) {
         super.renderLabels(poseStack, mouseX, mouseY);
         progressWidget.render(poseStack, mouseX - leftPos, mouseY - topPos, 0);
-    }
-
-    @Override
-    public void render(final PoseStack poseStack, final int mouseX, final int mouseY, final float delta) {
-        renderBackground(poseStack);
-        super.render(poseStack, mouseX, mouseY, delta);
-        renderTooltip(poseStack, mouseX, mouseY);
     }
 }
