+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/screen/ExporterScreen.java
@@ -5,11 +5,7 @@ import com.refinedmods.refinedstorage2.platform.common.containermenu.property.Pr
 import com.refinedmods.refinedstorage2.platform.common.screen.widget.ExactModeSideButtonWidget;
 import com.refinedmods.refinedstorage2.platform.common.screen.widget.ExporterSchedulingModeSideButtonWidget;
 import com.refinedmods.refinedstorage2.platform.common.screen.widget.RedstoneModeSideButtonWidget;
-import com.refinedmods.refinedstorage2.platform.common.screen.widget.ResourceFilterButtonWidget;
 
-import com.mojang.blaze3d.systems.RenderSystem;
-import com.mojang.blaze3d.vertex.PoseStack;
-import net.minecraft.client.renderer.GameRenderer;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.player.Inventory;
@@ -21,10 +17,6 @@ public class ExporterScreen extends AbstractBaseScreen<ExporterContainerMenu> {
 
     public ExporterScreen(final ExporterContainerMenu menu, final Inventory playerInventory, final Component text) {
         super(menu, playerInventory, text);
-
-        this.titleLabelX = 7;
-        this.titleLabelY = 7;
-        this.inventoryLabelX = 7;
         this.inventoryLabelY = 42;
         this.imageWidth = 210;
         this.imageHeight = 137;
@@ -46,33 +38,15 @@ public class ExporterScreen extends AbstractBaseScreen<ExporterContainerMenu> {
             getMenu().getProperty(PropertyTypes.EXPORTER_SCHEDULING_MODE),
             this::renderComponentTooltip
         ));
-
-        final ResourceFilterButtonWidget resourceFilterButton = new ResourceFilterButtonWidget(
-            leftPos + imageWidth - ResourceFilterButtonWidget.WIDTH - 7 - 34,
-            topPos + 4,
-            menu
-        );
-        addRenderableWidget(resourceFilterButton);
     }
 
     @Override
-    protected void renderBg(final PoseStack poseStack, final float delta, final int mouseX, final int mouseY) {
-        RenderSystem.setShader(GameRenderer::getPositionTexShader);
-        RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, 1.0F);
-        RenderSystem.setShaderTexture(0, TEXTURE);
-
-        final int x = (width - imageWidth) / 2;
-        final int y = (height - imageHeight) / 2;
-
-        blit(poseStack, x, y, 0, 0, imageWidth, imageHeight);
-
-        super.renderBg(poseStack, delta, mouseX, mouseY);
+    protected int getResourceFilterButtonX() {
+        return super.getResourceFilterButtonX() - 34;
     }
 
     @Override
-    public void render(final PoseStack poseStack, final int mouseX, final int mouseY, final float delta) {
-        renderBackground(poseStack);
-        super.render(poseStack, mouseX, mouseY, delta);
-        renderTooltip(poseStack, mouseX, mouseY);
+    protected ResourceLocation getTexture() {
+        return TEXTURE;
     }
 }
